connect api page detail course
